requirements.txt
flask==2.3.3
spotipy==2.23.0
confluent-kafka==2.3.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
requests==2.31.0
python-dotenv==1.0.0
ollama==0.1.7
Werkzeug==2.3.7